Add an Image Server source to a toro map TODO: WIP

Add an Image Server source to a toro map TODO: WIP

Usage

add_tiles_from_map_server(map, url, tile_id, as_image_layer = FALSE, ...)

Arguments

map

A toro map object or a map proxy object.

url

The URL of the ArcGIS Image Server.

...

Additional parameters for the Image Server source.

Value

The updated map object.