Get the colours for a column in a map layer based on step breaks.

Allows the data to be styled by the step breaks in the column.

Usage

get_column_step_colours(column_name, breaks, colours)

Arguments

column_name

String representing the name of the column to be used.

breaks

Numeric vector of thresholds (must be sorted ascending).

colours

Vector of colours, length = length(breaks) + 1.

Value

          List containing the paint or layout option to be set.

Examples

get_column_step_colours("value", c(10, 20, 30), c("red", "orange", "yellow", "green", "blue"))
#> Error in get_column_step_colours("value", c(10, 20, 30), c("red", "orange",     "yellow", "green", "blue")): length(colours) == length(breaks) + 1 is not TRUE