Get layout options for a specific layer type

This function returns a list of layout options based on the layer type and any additional options provided.

Usage

get_layout_options(layer_type, options = list())

Arguments

layer_type

A string indicating the type of layer (e.g., "fill", "circle", "line").

options

A list of additional options to customize the layout properties.

Value

      A list of layout options suitable for the specified layer type.

See also

Examples

get_layout_options("line", list(line_cap = "butt", line_join = "bevel"))
#> $`line-cap`
#> [1] "butt"
#> 
#> $`line-join`
#> [1] "bevel"
#> 
get_layout_options("symbol", list(icon_image = "yellow_pin", icon_size = 1.5))
#> $`icon-allow-overlap`
#> [1] TRUE
#> 
#> $`icon-image`
#> [1] "yellow_pin"
#> 
#> $`icon-size`
#> [1] 1.5
#> 
#> $`icon-anchor`
#> [1] "bottom"
#> 
#> $`icon-offset`
#> $`icon-offset`[[1]]
#> [1] 0
#> 
#> $`icon-offset`[[2]]
#> [1] 0
#> 
#> 
#> $`text-font`
#> $`text-font`[[1]]
#> [1] "Open Sans Regular"
#> 
#> 
#> $`text-size`
#> [1] 12
#>