Get paint options for a specific layer type

This function returns a list of paint options based on the layer type and any additional options provided.

Usage

get_paint_options(layer_type, options = list())

Arguments

layer_type

A string indicating the type of layer (e.g., "fill", "circle", "line").

options

A list of additional options to customize the paint properties.

Value

      A list of paint options suitable for the specified layer type.

See also

Examples

get_paint_options("line", list(colour = "blue", opacity = 0.8, line_width = 2))
#> $`line-color`
#> [1] "blue"
#> 
#> $`line-opacity`
#> [1] 0.8
#> 
#> $`line-width`
#> [1] 2
#> 
get_paint_options("circle", list(colour = "red", circle_radius = 10, outline_colour = "black"))
#> $`circle-color`
#> [1] "red"
#> 
#> $`circle-opacity`
#> [1] 1
#> 
#> $`circle-radius`
#> [1] 10
#> 
#> $`circle-stroke-color`
#> [1] "black"
#> 
#> $`circle-stroke-opacity`
#> [1] 1
#> 
#> $`circle-stroke-width`
#> [1] 1
#> 
# Use with get_column for data-driven styling:
get_paint_options("fill", list(colour = get_column("color"), opacity = get_column("opacity")))
#> $`fill-color`
#> $`fill-color`[[1]]
#> [1] "get"
#> 
#> $`fill-color`[[2]]
#> [1] "color"
#> 
#> 
#> $`fill-opacity`
#> $`fill-opacity`[[1]]
#> [1] "get"
#> 
#> $`fill-opacity`[[2]]
#> [1] "opacity"
#> 
#> 
#> $`fill-outline-color`
#> [1] "grey"
#> 
get_paint_options("fill", list(
   colour = get_column_group("group", c("A" = "green", "B" = "blue"))
))
#> $`fill-color`
#> $`fill-color`[[1]]
#> [1] "match"
#> 
#> $`fill-color`[[2]]
#> $`fill-color`[[2]][[1]]
#> [1] "get"
#> 
#> $`fill-color`[[2]][[2]]
#> [1] "group"
#> 
#> 
#> $`fill-color`[[3]]
#> [1] "A"
#> 
#> $`fill-color`[[4]]
#> [1] "green"
#> 
#> $`fill-color`[[5]]
#> [1] "B"
#> 
#> $`fill-color`[[6]]
#> [1] "blue"
#> 
#> $`fill-color`[[7]]
#> [1] "#cccccc"
#> 
#> 
#> $`fill-opacity`
#> [1] 1
#> 
#> $`fill-outline-color`
#> [1] "grey"
#> 
get_paint_options("fill", list(
   opacity = get_column_step_colours("percent", c(25, 75), c("red", "orange", "yellow"))
))
#> $`fill-color`
#> [1] "grey"
#> 
#> $`fill-opacity`
#> $`fill-opacity`[[1]]
#> [1] "step"
#> 
#> $`fill-opacity`[[2]]
#> $`fill-opacity`[[2]][[1]]
#> [1] "get"
#> 
#> $`fill-opacity`[[2]][[2]]
#> [1] "percent"
#> 
#> 
#> $`fill-opacity`[[3]]
#> [1] "red"
#> 
#> $`fill-opacity`[[4]]
#> [1] 25
#> 
#> $`fill-opacity`[[5]]
#> [1] "orange"
#> 
#> $`fill-opacity`[[6]]
#> [1] 75
#> 
#> $`fill-opacity`[[7]]
#> [1] "yellow"
#> 
#> 
#> $`fill-outline-color`
#> [1] "grey"
#>