Create a Maplibre map widget.

This function creates a map htmlwidget for use in R and Shiny applications.

Usage

map(
  style = "light-grey",
  center = c(174, -41),
  zoom = 2,
  width = "100%",
  height = NULL,
  session = shiny::getDefaultReactiveDomain(),
  ...
)

Arguments

style

The style of the map. Default is "light-grey".

center

The initial center of the map as a longitude/latitude pair. Default is c(174, -41).

zoom

The initial zoom level of the map. Default is 2.

width

The width of the widget. Optional.

height

The height of the widget. Optional.

session

The Shiny session object. Default is the current session.

...

Additional options to customize the map.

Value

     An object of class \code{htmlwidget} representing the gauge plot.

Examples

if (interactive()) {
  map()
}